### 2019-06-24 #317 @biagolini
GenTag: A package to improve animal color tagging protocol  
Submitting Author: Carlos Biagolini-Jr (@biagolini)
Repository: https://github.com/biagolini/GenTag
Version submitted: 1.0

I would like to submit the R package entitled 'GenTag' for consideration in rOpenSci. This is a package which proposes an improved protocol for animal color tagging. Since the early 1980s, several studies have suggested a negative influence of color tags on animal social behavior. However, methodological approaches to deal with this issue are lacking.  'GenTag' was designed to produce random sequences of color tag combinations. The package was designed to run on R base, without cross-package dependence. The computational routine is simple, implying low CPU usage. Despite the simplicity of the routine itself, the application of this method is exceptionally broad, as it can be applied in any project that uses visual identification of animals based on color marking, across a wide range of taxa, including birds, mammals, and fishes.
