### 2019-06-27 #317 @biagolini
Thank you for your attention, @noamross. I included: README, vignette, Travis and AppVeyor CI continuous integration; also I update the documentation to Roxygen-based files.

Best,
- Carlos Biagolini
