### 2019-06-27 #317 @biagolini
> Thank you for the inquiry, @biagolini. This package is a bit of an edge case topically. We don't have a package category specifically that it fits in to, but we think that packages like this - that are designed to facilitate field and laboratory processes and set up experimental designs, are in-scope, so we may say so explicitly in the next revision of our [Aims and Scope](https://devguide.ropensci.org/policies.html#aims-and-scope).
>
> However, I do note that this package is not yet ready for review. We require a number of things before sending a package to review, including a full README and vignette, a test suite with continuous integration and test coverage reporting, and Roxygen-based documentation. Please consult our [packaging guide](https://devguide.ropensci.org/building.html), which describes our packaging standards, in getting your package prepared for a full submission, and feel free to use our [forum](https://discuss.ropensci.org/) if you have questions.
