### 2019-06-28 #317 @noamross
Thank you @biagolini, fast work!  I note that while you have set up continuous integration and test coverage, you still do not have any tests for your package. As stated in our guide, we aim for at least 75% test coverage in the test suite, any less needs specific justification.

When you package is ready, please open a new issue using the full submission template.
