+++ R/combination_generators.R
@@ -1,0 +1,368 @@
+#' @title All equal tag sequence sample
+#' @description Create combinations with equal sample probability to all colors.
+#' @param ncombinations Number of combinations to be generated.
+#' @param ntag Number of tag to be used in each animal.
+#' @param colorsname Names/Code of color tags to be sample.
+#' @param gen_method method used for sample colors for tag sequence.
+#' @param usedcombinations Pre used combinations.
+#' @param colorsf Frequencies/ratio for color sample.
+#' @param nspecial Number of special tags/codes, such as metallic, 'EMPTY', or flag (min 0, max 2).
+#' @param name1 Name of special tag 1.
+#' @param name2 Name of special tag 2.
+#' @param location1 Position (or group of positions) to special band 1.
+#' @param location2 Position (or group of positions) to special band 2.
+#' @param nspecial1 Number of special tag 1 that will be present in all sequences genetated.
+#' @param nspecial2 Number of special tag 2 that will be present in all sequences genetated.
+#' @param emptyused If pre-used combination has code for empty set as TRUE, otherwise FALSE
+#' @param emptyname Code used to define empty code.
+#' @param currentyear Current year.
+#' @param yearsurvival An estimation of the proportion of animals that survive between years.
+#' @param lifespan Combinations older the lifespan will be automatically disregard.
+#' @param iotf Ignore older than lifespan. If TRUE ignore pre-used combinations older than lifespan
+#' @param yearusedcombinations The year in which the combination was used.
+#' @param speed Speed for color frequency adjustment.
+#' @param ignorecolor Color to be ignored on lifexp.
+#' @param g1 Group of tags from group 1. For example, in a code of 6 colors for bird tag, tags from 1 to 3 belongs to the left leg, and  tags from 1 to 3 belongs to the right. g1 must to address the left leg, and g2 must to address right leg. Thus: 'g1=1:3' and 'g2=4:6'.
+#' @param g2 Group of tags from group 2. For example, in a code of 6 colors for bird tag, tags from 1 to 3 belongs to the left leg, and  tags from 1 to 3 belongs to the right. g1 must to address the left leg, and g2 must to address right leg. Thus: 'g1=1:3' and 'g2=4:6'.
+#' @param g3 Group of tags from group 3.
+#' @param g4 Group of tags from group 4.
+#' @param g5 Group of tags from group 5.
+#' @param g6 Group of tags from group 6.
+#' @param parameterslist parameters for methods not provide by 'GenTag'
+#' @return A list of combinations
+#' @examples
+#' # Create an object contain the name/code of tag colors
+#' tcol<-c('Black','Blue','Brown','Gray','Green','Pink','Purple','Red','White','Yellow')
+#' # Generate color tag combination without especial tags
+#' genseq(30, 4, colorsname= tcol)
+#' # Generate color tag combination with especial color (ex metallic tag for numeric identification)
+#' genseq(30, 4, tcol, nspecial=1, name1='Metal', location1=c(2,4))
+#' # For ongoing works, use the argument usedcombinations to informe the previus used combinations
+#' data(pre_used) # Data example
+#' genseq(30, 4, colorsname= tcol, usedcombinations=pre_used[,1:4])
+#' combinations<-genseq(100, 4, tcol) # save combinations into an object
+#' @export
+genseq <- function(ncombinations = 100, ntag = 4, colorsname, gen_method = "allequal",
+                   usedcombinations = NA, colorsf = NA, nspecial = 0, name1 = "Metal",
+                   name2 = "EMPTY", location1 = 1, location2 = 2, nspecial1 = 1, nspecial2 = 1,
+                   emptyused = FALSE, emptyname = "EMPTY", currentyear = NA, yearsurvival = 1,
[…]
