### 2019-07-01 e9ee9a2
Ajustment in travis file
