+++ .travis.yml
@@ -1,0 +1,14 @@
+language: R
+cache: packages
+
+os:
+  - linux
+  - osx
+  
+r:
+  - 3.5
+  - oldrel
+  - release
+
+after_success:
+    - Rscript -e 'covr::codecov(type="all", token="9c054f8c-9476-4e2d-ab6d-f0cef953fbbf")'
