+++ DESCRIPTION
@@ -11,3 +11,6 @@ URL: https://github.com/biagolini/GenTag
 Depends: base
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
+Suggests: 
+    knitr,
+    rmarkdown
