### 2019-07-01 37b4b8d
fixing erro in synfunc
