+++ R/Imports.R
@@ -1,0 +1,2 @@
+#' # Functions to be imported
+#' @importFrom base stop warning numeric format Sys.time matrix cbind unique length which NA
