### 2019-07-02 d268780
add testthat
