+++ DESCRIPTION
@@ -15,4 +15,5 @@ RoxygenNote: 6.1.1
 Suggests: 
     knitr,
     rmarkdown,
-    covr
+    covr,
+    testthat (>= 2.1.0)
