### 2019-07-22 #317 @noamross
Hi @biagolini, I'm closing this issue.  Please so open a new one when you're ready to submit.  A quick note - when you test for coverage, it should be evaluated using`"tests"`, rather than `"all"` as is currently set in your `.travis.yml`.
